[Python] Lazy reading and writing. (#2878)
This refactors the read/write (and _load/_save) methods considerably and
introduces:

- `Model.read(toml_path, internal=False, external=False)` only reads the
model config, and nothing else. It means no tables, or empty ones in the
case of Node/Link, which can be written to disk as well. It's much like
the `Model(startime, endtime, crs)` constructor result in that sense.
- `Node.read(external=True, internal=True)` to load all node tables from
disk (overwriting/resetting them) if any changes were done.
- `TableModel.read()` for individual tables, same caveats as above.


- `Model.write(toml_path, toml=True, internal=True, external=True)` to
control writing all geopackage tables (internal) and/or all arrow/netcdf
files (external).
- `Node.write(internal=True, external=True)` for writing nodes (e.g. all
Basin tables)
- `TableModel.write()` to write individual tables

One needs to set the Model (toml) filepath first, so the input folder
can be determined.

This allows for more granular control of writing for the HydroMT plugin.
However, note that this API introduces some edge cases that need careful
handling. It becomes much easier to create invalid/inconsistent models.
For example, reading a single table and afterwards adding new nodes
might clash (non-unique node ids) with another yet unread table. That
case will be solved by #1813. Another one is writing a single
non-spatial table, yielding a sqlite database, not a valid geopackage.

This also refactors ChildModel and introduces a ParentModel (both now
using composition instead of inheritance) so an empty lazy table can
find the toml path of its (grandparent) Model.

@hboisgon see the `test_hydromt.py` for API examples that might
kickstart the HydroMT plugin.

---------

+++ python/ribasim/ribasim/model.py
@@ -56,8 +56,10 @@ from ribasim.input_base import (
     FileModel,
     ParentModel,
     SpatialTableModel,
+    _init_context_var,
     context_file_loading,
     context_file_writing,
+    init_context,
 )
 from ribasim.utils import (
     MissingOptionalModule,
@@ -78,7 +80,10 @@ logger = logging.getLogger(__name__)
 
 
 class Model(FileModel, ParentModel):
-    """A model of inland water resources systems."""
+    """The main class to represent a Ribasim model.
+
+    It represents the toml file almost directly (same fields).
+    """
 
     starttime: datetime.datetime
     endtime: datetime.datetime
@@ -119,6 +124,18 @@ class Model(FileModel, ParentModel):
     node: NodeTable = Field(default_factory=NodeTable)
     link: LinkTable = Field(default_factory=LinkTable)
     use_validation: bool = Field(default=True, exclude=True)
+
+    @classmethod
+    def allows_lazy(cls) -> bool:
+        """Whether this FileModel allows lazy loading."""
+        # We can't lazily load the Model itself as it needs
+        # to read the TOML file to determine other filepaths.
+        return False
+
+    @classmethod
+    def default_filepath(cls) -> Path:
+        """Return the default filepath for this FileModel."""
+        return Path("ribasim.toml")
 
     @field_validator("ribasim_version")
     @classmethod
@@ -200,7 +217,7 @@ class Model(FileModel, ParentModel):
         INDENT = "    "
         for field in self.model_fields_set:
             attr = getattr(self, field)
-            if isinstance(attr, LinkTable):
+            if isinstance(attr, (LinkTable, NodeTable)):
                 content.append(f"{INDENT}{field}=Link(...),")
             else:
                 if isinstance(attr, NodeModel):
@@ -225,39 +242,42 @@ class Model(FileModel, ParentModel):
         Path
             The file path of the written TOML file.
         """
-        content = self.model_dump(
-            exclude_unset=True, exclude_none=True, by_alias=True, context="write"
-        )
+        content = self.model_dump(exclude_unset=True, exclude_none=True, by_alias=True)
         # Filter empty dicts (default Nodes)
         content = dict(filter(lambda x: x[1], content.items()))
         with fn.open("wb") as f:
             tomli_w.dump(content, f)
         return fn
 
-    def _save(self, directory: DirectoryPath, input_dir: DirectoryPath):
+    def _write(
+        self,
+        directory: DirectoryPath,
+        internal: bool = True,
+        external: bool = True,
+    ) -> None:
         # We write all tables to a temporary GeoPackage with a dot prefix,
         # and at the end move this over the target file.
         # This does not throw a PermissionError if the file is open in QGIS.
-        db_path = directory / input_dir / ".database.gpkg"
 
-        # avoid adding tables to existing model
-        db_path.parent.mkdir(parents=True, exist_ok=True)
-        db_path.unlink(missing_ok=True)
-        context_file_writing.get()["database"] = db_path
+        if internal:
+            db_path = directory / ".database.gpkg"
 
-        self.link._save(directory, input_dir)
-        node = self.node
+            # avoid adding tables to existing model
+            db_path.unlink(missing_ok=True)
 
-        assert node.df is not None
-        node._save(directory, input_dir)
+            self.link.write()
+            node = self.node
+            assert node.df is not None
+            node.write()
 
-        # Run after geopackage schema has been created
-        _set_db_schema_version(db_path, ribasim.__schema_version__)
+            # Run after geopackage schema has been created
+            _set_db_schema_version(db_path, ribasim.__schema_version__)
 
         for sub in self._nodes():
-            sub._save(directory, input_dir)
+            sub.write(internal=internal, external=external)
 
-        shutil.move(db_path, db_path.with_name("database.gpkg"))
+        if internal:
+            shutil.move(db_path, db_path.with_name("database.gpkg"))
 
     def set_crs(self, crs: str) -> None:
         """Set the coordinate reference system of the data in the model.
@@ -327,19 +347,39 @@ class Model(FileModel, ParentModel):
                 yield attr
 
     @classmethod
-    def read(cls, filepath: str | PathLike[str]) -> "Model":
+    def read(
+        cls, filepath: str | PathLike[str], internal: bool = True, external: bool = True
+    ) -> "Model":
         """Read a model from a TOML file.
 
         Parameters
         ----------
         filepath : str | PathLike[str]
             The path to the TOML file.
+        internal : bool, optional
+            Load data from the database. Default is True.
+        external : bool, optional
+            Load data from the NetCDF input files. Default is True.
         """
         if not Path(filepath).is_file():
             raise FileNotFoundError(f"File '{filepath}' does not exist.")
-        return cls(filepath=filepath)  # type: ignore
 
-    def write(self, filepath: str | PathLike[str]) -> Path:
+        with init_context(
+            {
+                "internal": internal,
+                "external": external,
+                "directory": Path(filepath).parent,
+            }
+        ):
+            return cls(filepath=Path(filepath).name)  # type: ignore
+
+    def write(
+        self,
+        filepath: str | PathLike[str],
+        toml: bool = True,
+        internal: bool = True,
+        external: bool = True,
+    ) -> Path:
         """Write the contents of the model to disk and save it as a TOML configuration file.
 
         If ``filepath.parent`` does not exist, it is created before writing.
@@ -348,22 +388,32 @@ class Model(FileModel, ParentModel):
         ----------
         filepath : str | PathLike[str]
             A file path with .toml extension.
+        toml : bool, optional
+            Write the TOML configuration file. Default is True.
+        internal : bool, optional
+            Write the database. Default is True.
+        external : bool, optional
+            Write the NetCDF input files. Default is True.
         """
         if self.use_validation:
             self._validate_model()
 
         filepath = Path(filepath)
-        self.filepath = filepath
         if not filepath.suffix == ".toml":
             raise ValueError(f"Filepath '{filepath}' is not a .toml file.")
-        context_file_writing.set({})
-        directory = filepath.parent
-        directory.mkdir(parents=True, exist_ok=True)
-        self._save(directory, self.input_dir)
-        fn = self._write_toml(filepath)
+        self.filepath = filepath
 
         context_file_writing.set({})
-        return fn
+        directory = filepath.parent / self.input_dir
+        directory.mkdir(parents=True, exist_ok=True)
+        context_file_writing.get()["directory"] = directory
+
+        self._write(directory, internal=internal, external=external)
+        if toml:
+            self._write_toml(filepath)
+
+        context_file_writing.set({})
+        return filepath
 
     def _validate_model(self):
         df_link = self.link.df
@@ -487,13 +537,14 @@ class Model(FileModel, ParentModel):
             with filepath.open("rb") as f:
                 config = tomli.load(f)
 
+            config["filepath"] = filepath
             directory = filepath.parent / config["input_dir"]
             context_file_loading.get()["directory"] = directory
-            db_path = directory / "database.gpkg"
+            _init_context_var.get()["directory"] = directory  # type: ignore
 
+            db_path = directory / "database.gpkg"
             if not db_path.is_file():
                 raise FileNotFoundError(f"Database file '{db_path}' does not exist.")
-
             context_file_loading.get()["database"] = db_path
 
             return config
@@ -505,6 +556,20 @@ class Model(FileModel, ParentModel):
         # Drop database info
         context_file_loading.set({})
         return self
+
+    @property
+    def database_path(self) -> FilePath | None:
+        """
+        Get the path to the database file.
+
+        Returns
+        -------
+        FilePath | None
+            The path to the database file.
+        """
+        if self.filepath:
+            return self.filepath.parent / self.input_dir / "database.gpkg"
+        return None
 
     @property
     def toml_path(self) -> FilePath:
